+++ drivers/gpu/nvgpu/common/pmu/pmu_gp10b.c
@@ -265,7 +265,7 @@ void gp10b_pmu_elpg_statistics(struct gk20a *g, u32 pg_engine_id,
 
 	nvgpu_falcon_copy_from_dmem(pmu->flcn,
 		pmu->stat_dmem_offset[pg_engine_id],
-		(u8 *)&stats, sizeof(struct pmu_pg_stats_v1), 0);
+		(u8 *)&stats, (u32)sizeof(struct pmu_pg_stats_v1), 0);
 
 	pg_stat_data->ingating_time = stats.total_sleep_timeus;
 	pg_stat_data->ungating_time = stats.total_nonsleep_timeus;
